# backend-team-final-project-sanbercode-2022/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
        "/CartItems/{id}": {
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Update CartItems by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CartItems"
                ],
                "summary": "Update CartItems.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CartItems id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CartItems"
                        }
                    }
                }
            }
        },
...
        "/cartItems": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Get a list of CartItems.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CartItems"
                ],
                "summary": "Get all cartItem.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CartItems"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Creating a new CartItems.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CartItems"
                ],
                "summary": "Create New CartItems.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CartItems"
                        }
                    }
                }
            }
        },
        "/cartItems/{id}": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Get an CartItems by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CartItems"
                ],
                "summary": "Get CartItems.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CartItems ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CartItems"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Delete a Cart item by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CartItems"
                ],
                "summary": "Delete one Cart item.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CartItem id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CartItems": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "projects": {
                    "$ref": "#/definitions/models.Projects"
                },
                "projects_id": {
                    "description": "Relation",
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "shopping_cart": {
                    "$ref": "#/definitions/models.ShoppingCart"
                },
                "shopping_cart_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Category": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "projects": {
                    "description": "Relation",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Projects"
                    }
                }
            }
        },
        "models.Images": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "images_url": {
                    "type": "string"
                },
                "projects": {
                    "$ref": "#/definitions/models.Projects"
                },
                "projects_id": {
                    "description": "Relation",
                    "type": "string"
                }
            }
        },
        "models.Installment": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "projects": {
                    "$ref": "#/definitions/models.Projects"
                },
                "projects_id": {
                    "description": "Relation",
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.Projects": {
            "type": "object",
            "properties": {
                "cart_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CartItems"
                    }
                },
                "category": {
                    "$ref": "#/definitions/models.Category"
                },
                "category_id": {
                    "description": "Relation",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dueDate": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Images"
                    }
                },
                "installment": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Installment"
                    }
                },
                "margin": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "periode": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.ShoppingCart": {
            "type": "object",
            "properties": {
                "cartItems": {
                    "description": "Relation",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CartItems"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "payment_status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.Transaction": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "credit": {
                    "type": "integer"
                },
                "debit": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "installment_id": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "sender": {
                    "type": "string"
                },
                "shopping_cart_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "user_id": {
                    "description": "Relation",
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3
                },
                "id": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 6
                },
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Projects"
                    }
                },
                "role": {
                    "description": "3 roles : admin, investor, investee",
                    "type": "string"
                },
                "shoppingCart": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ShoppingCart"
                    }
                },
                "status": {
                    "type": "string"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Transaction"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "userProfile": {
                    "description": "Relation",
                    "$ref": "#/definitions/models.UserProfile"
                }
            }
        },
        "models.UserProfile": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "bank_account_number": {
                    "type": "integer"
                },
                "bank_name": {
                    "type": "string"
                },
                "city": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "created_at": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "gender": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "id": {
                    "type": "string"
                },
                "ktp": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5
                },
                "profile_url": {
                    "type": "string"
                },
                "province": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}